--- docs/shared_parsers_catalog/grub2_editenv_list.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.parsers.grub2_editenv_list
-   :members:
-   :show-inheritance:
