+++ insights/parsers/grubenv.py
@@ -1,7 +1,6 @@
 """
 GrubEnv - file ``/boot/grub2/grubenv``
 ======================================
-
 This parser reads the GRUB environment block file. The file is laid out in a
 key=value format similar to an ini file except it doesn't have any headers.
 
@@ -9,6 +8,11 @@ The parser stores the key/value pairs in itself which inherits a dict. This
 file is only used in Grub 2, but in RHEL8 with BLS being the default. There
 were several variables added that are referenced in the
 ``/boot/loader/entries/*.conf`` files.
+
+Grub2EditenvList - command ``grub2-editenv list``
+=================================================
+This module parses the output of the command ``grub2-editenv list``, which
+list the current variables in GRUB environment block file or error message.
 """
 from insights import get_active_lines, parser, Parser
 from insights.parsers import SkipException
@@ -90,3 +94,61 @@ class GrubEnv(Parser, dict):
     @property
     def tuned_params(self):
         return self.get("tuned_params", "")
+
+
+@parser(Specs.grubenv)
+class Grub2EditenvList(GrubEnv):
+    """
+    Parses the ``grub2-editenv list`` command and returns a dict
+    of the grubenv variables. The command output without error
+    message is same as the content in ``/boot/grub2/grubenv``.
+
+    Sample output of the command::
+
+        saved_entry=295e1ba1696e4fad9e062f096f92d147-4.18.0-305.el8.x86_64
+        kernelopts=root=/dev/mapper/root_vg-lv_root ro crashkernel=auto resume=/dev/mapper/root_vg-lv_swap rd.lvm.lv=root_vg/lv_root rd.lvm.lv=root_vg/lv_swap console=tty0 console=ttyS0,115200
+        boot_success=0
+        boot_indeterminate=2
+        tuned_params=transparent_hugepages=never
+        tuned_initrd=
+
+    Examples:
+
+        >>> type(grub2_editenv_list)
+        <class 'insights.parsers.grubenv.Grub2EditenvList'>
+        >>> grub2_editenv_list.has_kernelopts
+        True
+        >>> grub2_editenv_list.kernelopts
+        'root=/dev/mapper/root_vg-lv_root ro crashkernel=auto resume=/dev/mapper/root_vg-lv_swap rd.lvm.lv=root_vg/lv_root rd.lvm.lv=root_vg/lv_swap console=tty0 console=ttyS0,115200'
+        >>> grub2_editenv_list.has_tuned_params
+        True
+        >>> grub2_editenv_list.tuned_params
+        'transparent_hugepages=never'
+        >>> grub2_editenv_list['saved_entry']
+        '295e1ba1696e4fad9e062f096f92d147-4.18.0-305.el8.x86_64'
+    """
+
+    def __init__(self, context):
+        super(Grub2EditenvList, self).__init__(context)
+
+    def parse_content(self, content):
+        if not content:
+            raise SkipException("Empty output.")
+
+        self.error = []
+        data = dict()
+        for line in get_active_lines(content):
+            if "=" not in line:
+                self.error.append(line)
+                continue
+
+            key, value = line.split("=", 1)
+            if not value:
+                continue
+
+            data[key] = value
+
+        if (not data) and (not self.error):
+            raise SkipException("No parsed data.")
+
+        self.update(data)
